Feat: Split trellogithub.go up into smaller ones
